Move emailLogin tests to a separate module and refactor
